Add unit tests and fixes for APSTUD-3293 - Editors Do Not Retain Text Selection After Block Commenting
I took the title literally and had only fixed the behavior for block comment toggling. The issue is that we didn't retain selections after commenting blocks of code with line prefix comments. I added assertions to the unit tests to verify the proper selection behavior and then fixed the code.
  - if multiple lines are selected, we retain the selection after toggling
  - if there is no selection for line comment toggling, we retain the caret's column in the text. (i.e. '# Com|ment' -> 'Com|ment')
  - if there's no selection and the caret is at column 0, we select the inserted comment prefix to match JDT behavior.